pyproject.toml
[project]
name = "site"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "cryptography>=46.0.6",
    "daphne>=4.2.1",
    "django>=6.0.3",
    "django-cleanup>=9.0.0",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.8",
    "ipython>=9.12.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
